Moved CoreValidator back to original package
