+++ core-plugins/src/test/java/co/cask/hydrator/plugin/transform/StructuredRecordToGenericRecordTransformTest.java
@@ -23,6 +23,7 @@ import co.cask.cdap.etl.api.TransformContext;
 import co.cask.cdap.etl.mock.common.MockEmitter;
 import co.cask.cdap.etl.mock.common.MockPipelineConfigurer;
 import co.cask.cdap.etl.mock.transform.MockTransformContext;
+import co.cask.hydrator.plugin.validator.CoreValidator;
 import com.google.common.collect.ImmutableMap;
 import org.apache.avro.generic.GenericRecord;
 import org.junit.Assert;
