+++ core-plugins/src/test/java/co/cask/hydrator/plugin/transform/ValidatorTransformTest.java
@@ -22,6 +22,7 @@ import co.cask.cdap.etl.api.Validator;
 import co.cask.cdap.etl.mock.common.MockEmitter;
 import co.cask.cdap.etl.mock.common.MockPipelineConfigurer;
 import co.cask.cdap.etl.mock.transform.MockTransformContext;
+import co.cask.hydrator.plugin.validator.CoreValidator;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import org.junit.Assert;
